[작업중] 출근 전 저장...?
뒤늦게 발견한 머지 충돌은 무시합니다.

+++ HeadingNet/Inc/Base/HeadingBuffer.h
@@ -22,7 +22,7 @@ namespace Heading
 		void	get_send_data	( char**		_buffer
 								, int*			_length		);
 		bool	set_data		( char* _data
-								, uint64_t _length );
+								, int _length				);
 
 		bool	isEmpty			();
